refactor(kernel::syscalls): and also some docs
